Main classes done, just figuring out the interface now. Current problem is how to track all the goals and update them when an event is recorded.

+++ prove/Develop05/BaseGoal.cs
@@ -8,7 +8,7 @@ class BaseGoal
     protected double _maxCompletions;
     protected bool _completed;
 
-    public BaseGoal(int pointsValue, string goalName, string goalDescription, int maxCompletions)
+    public BaseGoal(int pointsValue, string goalName, string goalDescription, double maxCompletions)
     {
         _pointsValue = pointsValue;
         _goalName = goalName;
