+++ prove/Develop05/CheckGoal.cs
@@ -1,7 +1,7 @@
 class CheckGoal : BaseGoal
 {
     private int _bonusPoints;
-    public CheckGoal(int pointsValue, string goalName, string goalDescription, int maxCompletions, int bonusPoints) :
+    public CheckGoal(int pointsValue, string goalName, string goalDescription, double maxCompletions, int bonusPoints) :
     base(pointsValue, goalName, goalDescription, maxCompletions)
     {
         _bonusPoints = bonusPoints;
