+++ prove/Develop05/SimpleGoal.cs
@@ -1,6 +1,6 @@
 class SimpleGoal : BaseGoal
 {
-    public SimpleGoal(int pointsValue, string goalName, string goalDescription, int maxCompletions) :
+    public SimpleGoal(int pointsValue, string goalName, string goalDescription, double maxCompletions) :
     base(pointsValue, goalName, goalDescription, maxCompletions)
     {
         SetMaxCompletions(1);
